airplay: Move StaticFileWebServer to http
Relates to #1078
